requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
Werkzeug==1.0.1
